requirements.txt
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
idna==2.7
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.2
pandas==0.23.4
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
urllib3==1.23
pillow==5.3